Add component for queuing messages.
All incomming messages are kept in memory. As soon as the queu reaches the
pre-configured size of 100 messages, the current list of messages is send.

You can reconfigure the queu size of the component by sending a message to
the `size` input port of this component. If the incomming data is not a
positive integer, an error message will be send to the `err` out port of this
component.